move parse into its own module
